Architectural refactoring of SendTransactionAsync - work in progress
