--- src/Stratis.Bitcoin/Broadcasting/Success.cs
@@ -1,9 +1,0 @@
-﻿namespace Stratis.Bitcoin.Broadcasting
-{
-    public enum Success
-    {
-        Yes,
-        DontKnow,
-        No
-    }
-}
