+++ src/Stratis.Bitcoin/Interfaces/IBroadcastManager.cs
@@ -7,7 +7,7 @@ namespace Stratis.Bitcoin.Interfaces
 {
     public interface IBroadcastManager
     {
-        Task<Success> TryBroadcastAsync(Transaction transaction);
+        Task<bool> TryBroadcastAsync(Transaction transaction);
 
         event EventHandler<TransactionBroadcastEntry> TransactionStateChanged;
